+++ lambda/postSchedulePeerTeachers/Scheduler/src/main/java/scheduler/Handler.java
@@ -196,7 +196,7 @@ public class Handler implements RequestHandler<APIGatewayProxyRequestEvent, APIG
 
         while (rs.next())
         {
-            people.put(rs.getInt("person_id"), new Person(rs.getInt("person_id"), rs.getString("email"), rs.getString("first_name"), rs.getString("last_name"), rs.getInt("desired_number_assignments"), rs.getBoolean("peer_teacher"), rs.getBoolean("teaching_assistant"), rs.getBoolean("administrator"), rs.getBoolean("professor")));
+            people.put(rs.getInt("person_id"), new Person(rs.getInt("person_id"), rs.getString("email"), rs.getString("first_name"), rs.getString("last_name"), rs.getFloat("desired_number_assignments"), rs.getBoolean("peer_teacher"), rs.getBoolean("teaching_assistant"), rs.getBoolean("administrator"), rs.getBoolean("professor")));
         }
 
         rs.close();
